feat: implement a (properly forked) watcher for nest
This is dope.
So now nest will spawn a watcher that watches the server and checks for
attached sessions. If none are attached, it will automatically cleanup
the unused session so as to prevent an unneeded server from running in
the background at all times for no reason.

If there is already a server, no watcher will be spawned, and instead
tmux will attach to the server with a new session.